requirements.txt
bcrypt==3.1.4
beautifulsoup4==4.5.3
bleach==1.5.0
cffi==1.11.2
defusedxml==0.5.0
Django==1.10.5
django-allauth==0.33.0
django-appconf==0.6
django-autoslug==1.9.3
django-cleanup==1.0.1
django-debug-toolbar==1.6
django-extra-fields==0.9
django-filter==1.0.2
django-froala-editor==2.7.1
django-html-sanitizer==0.1.5
django-jsonfield==0.9.13
django-mptt==0.8.7
django-pagedown==0.1.1
django-postman==3.1.0
django-render-block==0.5
django-summernote==0.2.3
#-e git://github.com/AASHE/django-tagulous@4f7dd6c04be0c70ccb5e23ca5652815bf3bb0258#egg=django_tagulous
django-templated-email==2.2.0
django-tinymce==2.6.0
django-webpack-loader==0.4.1
djangorestframework==3.6.2
djangorestframework-jwt==1.10.0
djangorestframework-recursive==0.1.1
docutils==0.14
easy-thumbnails==1.3
googlemaps==2.4.6
html5lib==0.9999999
lorem-ipsum-generator==0.3
mpttadmin==0.3.1
oauthlib==2.0.6
olefile==0.44
Pillow==4.0.0
#pkg-resources==0.0.0
psycopg2==2.7.3.2
pycparser==2.18
PyJWT==1.5.3
python-dateutil==2.6.1
python3-openid==3.1.0
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
sorl-thumbnail==12.4.1
sorl-thumbnail-serializer-field==0.2.1
sqlparse==0.2.2
transliterate==1.9
Unidecode==0.4.21
steem-piston
